refactured backend, broadcast merge to all users, fixed merge bug
